refactor kafka event mappers tests structure
